[FIX] updates to template method

+++ src/tests/class_tests/openms/source/MRMFeatureFilter_test.cpp
@@ -69,7 +69,7 @@ START_SECTION(bool checkRange(T value, T value_l, T value_u))
 {
   MRMFeatureFilter mrmff;
   // tests
-  TEST_EQUAL(mrmff.checkRange (2.0, 1.0, 2.0), true);
+  TEST_EQUAL(mrmff.checkRange((double)2.0, (double)1.0, (double)2.0), true);
   TEST_EQUAL(mrmff.checkRange(0.0, 1.0, 2.0), false);
   TEST_EQUAL(mrmff.checkRange(3.0, 1.0, 2.0), false);
   TEST_EQUAL(mrmff.checkRange(2, 1, 2), true);
